Integration Complete
Integration of ContactBook API and  contacts frontend is complete

+++ src/components/ContactCard.js
@@ -2,7 +2,6 @@ import React, { useContext } from 'react'
 import { useNavigate } from "react-router-dom";
 import { MdDelete, MdOutlineModeEditOutline } from "react-icons/md";
 import { removecontacts, getcontacts } from '../context/contactState';
-// import store from "../store/store"
 import contactContext from '../context/contactcontext';
 
 export default function ContactCard(props) {
@@ -10,7 +9,6 @@ export default function ContactCard(props) {
   const { deleteContact} = context
 
   let contact = props.contact
- 
   let mobile = contact.mobileNumbers[0].mobile
   let photo = `https://ui-avatars.com/api/?name=${contact.firstName}&length=1&background=random&size=262`
   const navigate = useNavigate();
@@ -18,6 +16,7 @@ export default function ContactCard(props) {
     <tr
       onClick={() => {
         navigate(`/contact/?id=${contact.id}&name=${contact.firstName}`);
+        localStorage.setItem('name', contact.firstName)
       }}
       role="button"
       className="hover:bg-[#f5f5f5] group relative"
@@ -30,24 +29,12 @@ export default function ContactCard(props) {
 
       <div className="hidden group-hover:visible invisible absolute top-0 right-4 h-full gap-3 sm:flex items-center">
         <button
-          className="h-full"
           onClick={(e) => {
             // Prevent parent handlers from firing
             e.stopPropagation();
-            navigate(`/contact/${contact.id}/edit`);
-          }}
-        >
-          <MdOutlineModeEditOutline
-            className="text-[#00000051] hover:text-black"
-            size={24}
-          />
-        </button>
-        <button
-          onClick={(e) => {
-            // Prevent parent handlers from firing
-            e.stopPropagation();
-            // if (confirm("Delete contact?")) {
-            deleteContact(contact.id);
+              if (window.confirm("Delete contact?")) {
+                deleteContact(contact.id);
+              }
             // }
           }}
         >
